+++ add_item.sh
@@ -4,8 +4,10 @@
 
 . lib/env.sh
 
-URL=$(echo "${*}" | jq -r .url)
-SITE=$(echo "${URL}" | cut -d/ -f3)
+log "add_item"
+
+URL=$(jq -j .url <<< "${*}")
+SITE=$(cut -d/ -f3 <<< "${URL}")
 
 P='return text returned of (display dialog'
 I='with icon caution'
@@ -13,32 +15,32 @@ T='with title "Add item"'
 
 # Get name
 CMD="${P} \"Enter name of new item\" ${I} ${T} default answer \"${SITE}\")"
-NAME=$(osascript -e "${CMD}")
+NAME=$(2>&- osascript -e "${CMD}")
 
 [ "${NAME}" == "" ] && exit
 
 # Get username
-SITE=$(echo "${NAME}" | cut -d/ -f3)
+SITE=$(cut -d/ -f3 <<< "${NAME}")
 
 CMD="${P} \"Enter username for ${SITE}\" ${I} ${T} default answer \"${bwuser}\")"
-USERNAME=$(osascript -e "${CMD}")
+USERNAME=$(2>&- osascript -e "${CMD}")
 
 [ "${USERNAME}" == "" ] && exit
 
 # Get password
-GENERATED=$(curl -s "${API}/generate?length=20&uppercase&lowercase&number&special" | jq -r .data.data)
+GENERATED=$(curl -s "${API}/generate?length=20&uppercase&lowercase&number&special" | jq -j .data.data)
 
 CMD="${P} \"Enter password for ${USERNAME}\" ${I} ${T} default answer \"${GENERATED}\" with hidden answer)"
-PASSWORD=$(osascript -e "${CMD}")
+PASSWORD=$(2>&- osascript -e "${CMD}")
 
 [ "${PASSWORD}" == "" ] && exit
 
 if [ "${PASSWORD}" != "${GENERATED}" ]; then
     CMD="${P} \"Confirm password for ${USERNAME}\" ${I} ${T} default answer \"\" with hidden answer)"
-    CONFIRM=$(osascript -e "${CMD}")
+    CONFIRM=$(2>&- osascript -e "${CMD}")
 
     if [ "${CONFIRM}" != "${PASSWORD}" ]; then
-	echo 'Add item,Passwords do not match'
+	echo -n 'Add item,Passwords do not match'
 	exit
     fi
 fi
@@ -60,6 +62,6 @@ PAYLOAD+=',"uris": [ { "match": null, "uri": "'"${URL}"'" }'
 PAYLOAD+='] } }'
 
 # Add item
-S=$(curl -s -H 'Content-Type: application/json' -d "${PAYLOAD}" "${API}"/object/item | jq -r .success)
+S=$(curl -s -H 'Content-Type: application/json' -d "${PAYLOAD}" "${API}"/object/item | jq -j .success)
 
 echo -n "Added ${SITE//,/ },Success: ${S}"