+++ lock.sh
@@ -4,5 +4,5 @@
 
 log "lock"
 
-curl -s -X POST "${API}"/lock | jq -r '.message // .data.title'
+curl -s -X POST "${API}"/lock | jq -j '.message // .data.title'
 rm -f "${DATA_DIR}"/*