+++ rm.sh
@@ -9,4 +9,4 @@ log "rm ${id} ${name}"
 curl -s -X DELETE "${API}"/object/item/"${id}" > /dev/null
 saveSync
 
-echo "${name}"
+echo -n "${name}"