+++ show_item.sh
@@ -11,7 +11,7 @@ export ITEM=$(curl -s "${API}/object/${field}/${id}")
 
 # Get item' type
 ICON="${PWD}/icon.png"
-TYPE=$(jq -r .data.type <<< "${ITEM}")
+TYPE=$(jq -j .data.type <<< "${ITEM}")
 
 case ${TYPE} in
     1)
@@ -24,7 +24,7 @@ case ${TYPE} in
 	;;
     3)
 	JQ="card"
-	BRAND=$(jq -L jq -r 'include "bw"; icon(.)' <<< "${ITEM}")
+	BRAND=$(jq -L jq -j 'include "bw"; icon(.)' <<< "${ITEM}")
 	ICON="${PWD}/icons/${BRAND}"
 	;;
     4)
@@ -36,16 +36,16 @@ esac
 log "Type = ${JQ}"
 
 # Get item's organization
-ORG_ID=$(jq -r .data.organizationId <<< "${ITEM}")
-ORG=$(jq -r --arg org "${ORG_ID}" '.data.data[] | select(.id == $org) | .name' "${DATA_DIR}"/organizations)
+ORG_ID=$(jq -j .data.organizationId <<< "${ITEM}")
+ORG=$(jq -j --arg org "${ORG_ID}" '.data.data[] | select(.id == $org) | .name' "${DATA_DIR}"/organizations)
 
 log "Org = ${ORG}"
 
 # Format item
-export DATA=$(jq -L jq -r --arg org "${ORG}" -f "jq/show_${JQ}.jq" <<< "${ITEM}")
+export DATA=$(jq -L jq -j --arg org "${ORG}" -f "jq/show_${JQ}.jq" <<< "${ITEM}")
 
 # Display dialog
-osascript \
+2>&- osascript \
     -e 'set t to (system attribute "DATA")' \
     -e 'set i to "'"${ICON}"'"' \
     -e 'display dialog t buttons {"OK"} default button "OK" with title "'"${alfred_workflow_name}"'" with icon posix file i' \